Fix name of transcode output
