+++ app/util/CredentialsGenerator.scala
@@ -1,12 +1,12 @@
 package util
 
 import com.amazonaws.services.securitytoken.model.AssumeRoleRequest
-import com.gu.media.aws.UploadAccess
+import com.gu.media.aws.{PipelineAccess, UploadAccess}
 import com.gu.media.logging.Logging
 import com.gu.media.upload.model.UploadCredentials
 import play.api.libs.json.{JsArray, JsObject, JsString, Json}
 
-class CredentialsGenerator(aws: UploadAccess) extends Logging {
+class CredentialsGenerator(aws: PipelineAccess) extends Logging {
   def forKey(key: String): UploadCredentials = {
     val keyPolicy = generateKeyPolicy(key)
