+++ common/src/main/scala/com/gu/media/upload/UploadKeys.scala
@@ -1,6 +1,6 @@
 package com.gu.media.upload
 
-import java.net.URLEncoder
+import java.time.format.DateTimeFormatter
 import java.time.{ZoneOffset, ZonedDateTime}
 
 case class UploadKey(folder: String, id: String) {
@@ -15,13 +15,15 @@ case class CompleteUploadKey(folder: String, id: String) {
   override def toString = s"$folder/$id/complete"
 }
 
-case class TranscoderOutputKey(year: Int, month: Int, day: Int, title: String, id: String, extension: String) {
-  override def toString = s"$year/$month/$day/${URLEncoder.encode(title, "UTF-8")}-$id.$extension"
+case class TranscoderOutputKey(prefix: String, title: String, id: String, extension: String) {
+  override def toString = s"$prefix/$title--$id.$extension"
 }
 
 object TranscoderOutputKey {
   def apply(title: String, id: String, extension: String): TranscoderOutputKey = {
     val now = ZonedDateTime.now(ZoneOffset.UTC)
-    TranscoderOutputKey(now.getYear, now.getMonthValue, now.getDayOfMonth, title, id, extension)
+    val prefix = now.format(DateTimeFormatter.ofPattern("yyyy/mm/dd"))
+
+    TranscoderOutputKey(prefix, title, id, extension)
   }
 }
